refactor(commands): split common CLI extension modules
