+++ src/codex_plugin_scanner/guard/runtime/command_common_cli_rules_1.py
@@ -1,0 +1,159 @@
+"""Common CLI command rules, batch 1."""
+
+from __future__ import annotations
+
+from .command_common_cli_matchers import (
+    _FIREBASE_DEPLOY,
+    _GLAB_DESTRUCTIVE,
+    _GLAB_MERGE,
+    _GLAB_VARIABLE,
+    _PRISMA_DEPLOY,
+    _PRISMA_DESTRUCTIVE,
+    _VAULT_ADMIN,
+    _VAULT_MUTATION,
+    _VAULT_READ,
+    _WRANGLER_CHANGE,
+    _WRANGLER_DESTRUCTIVE,
+)
+from .command_common_cli_rule_support import help_variants, rule
+from .command_rules import CommandSafetyRule
+
+COMMON_CLI_COMMAND_RULES_1: tuple[CommandSafetyRule, ...] = (
+    rule(
+            extension_id="command.platform.cloudflare",
+            suffix="destructive",
+            title="Cloudflare resource deletion",
+            description="Identifies Wrangler operations that permanently remove Cloudflare resources.",
+            matcher=_WRANGLER_DESTRUCTIVE,
+            action_class="Cloudflare destructive command",
+            risk_classes=("destructive_shell", "network_egress"),
+            safer_alternative="Inspect the exact account, environment, and resource before deletion.",
+            severity="critical",
+            safe_variants=help_variants(_WRANGLER_DESTRUCTIVE, short=True),
+            example_command="wrangler d1 delete production-db",
+        ),
+    rule(
+            extension_id="command.platform.cloudflare",
+            suffix="remote-change",
+            title="Cloudflare deployment or secret change",
+            description="Identifies Wrangler deployment and secret mutation operations.",
+            matcher=_WRANGLER_CHANGE,
+            action_class="Cloudflare production command",
+            risk_classes=("execution", "network_egress"),
+            safer_alternative="Review the target environment, deployment, and secret names before publishing.",
+            safe_variants=help_variants(_WRANGLER_CHANGE, short=True),
+            example_command="wrangler deploy",
+        ),
+    rule(
+            extension_id="command.gitlab",
+            suffix="destructive",
+            title="GitLab destructive administration",
+            description="Identifies project, release, or variable deletion through glab.",
+            matcher=_GLAB_DESTRUCTIVE,
+            action_class="GitLab destructive command",
+            risk_classes=("destructive_shell", "network_egress"),
+            safer_alternative="Inspect the exact GitLab project and resource before deletion.",
+            severity="critical",
+            safe_variants=help_variants(_GLAB_DESTRUCTIVE, short=True),
+            example_command="glab repo delete group/project",
+        ),
+    rule(
+            extension_id="command.gitlab",
+            suffix="merge",
+            title="GitLab merge request mutation",
+            description="Identifies merge or rebase operations that change remote GitLab repository state.",
+            matcher=_GLAB_MERGE,
+            action_class="GitLab merge command",
+            risk_classes=("execution", "network_egress"),
+            safer_alternative="Review the head, target branch, and pipeline state before merging or rebasing.",
+            safe_variants=help_variants(_GLAB_MERGE, short=True),
+            example_command="glab mr merge 42",
+        ),
+    rule(
+            extension_id="command.gitlab",
+            suffix="variable",
+            title="GitLab CI/CD variable access",
+            description="Identifies reads or writes of GitLab variables that may contain credentials.",
+            matcher=_GLAB_VARIABLE,
+            action_class="GitLab variable command",
+            risk_classes=("local_secret_read", "network_egress"),
+            safer_alternative="Use only the required variable and confirm the exact project or group scope.",
+            safe_variants=help_variants(_GLAB_VARIABLE, short=True),
+            example_command="glab variable get DEPLOY_TOKEN",
+        ),
+    rule(
+            extension_id="command.secrets.vault",
+            suffix="secret-read",
+            title="Vault secret read",
+            description="Identifies Vault reads that can return secret material to the invoking agent.",
+            matcher=_VAULT_READ,
+            action_class="Vault secret read command",
+            risk_classes=("local_secret_read", "network_egress"),
+            safer_alternative="Request only the specific secret path and fields required for the task.",
+            safe_variants=help_variants(_VAULT_READ, short=True),
+            example_command="vault kv get secret/app",
+        ),
+    rule(
+            extension_id="command.secrets.vault",
+            suffix="secret-mutation",
+            title="Vault secret mutation",
+            description="Identifies Vault operations that write, delete, destroy, or roll back secret data.",
+            matcher=_VAULT_MUTATION,
+            action_class="Vault secret mutation command",
+            risk_classes=("destructive_shell", "network_egress"),
+            safer_alternative="Confirm the mount, path, version, and recovery plan before changing secret state.",
+            severity="critical",
+            safe_variants=help_variants(_VAULT_MUTATION, short=True),
+            example_command="vault kv destroy -versions=3 secret/app",
+        ),
+    rule(
+            extension_id="command.secrets.vault",
+            suffix="security-administration",
+            title="Vault security administration",
+            description="Identifies token revocation and disabling of policies, auth methods, or secrets engines.",
+            matcher=_VAULT_ADMIN,
+            action_class="Vault security administration command",
+            risk_classes=("destructive_shell", "network_egress"),
+            safer_alternative="Inspect active leases, policies, and mount dependencies before disabling them.",
+            severity="critical",
+            safe_variants=help_variants(_VAULT_ADMIN, short=True),
+            example_command="vault token revoke accessor-value",
+        ),
+    rule(
+            extension_id="command.database.prisma",
+            suffix="destructive",
+            title="Prisma destructive database operation",
+            description="Identifies resets, direct SQL execution, and data-loss-accepted schema pushes.",
+            matcher=_PRISMA_DESTRUCTIVE,
+            action_class="Prisma destructive command",
+            risk_classes=("destructive_shell", "network_egress", "execution"),
+            safer_alternative="Inspect the schema diff and take a current backup before destructive Prisma operations.",
+            severity="critical",
+            safe_variants=help_variants(_PRISMA_DESTRUCTIVE, short=True),
+            example_command="npx prisma migrate reset",
+        ),
+    rule(
+            extension_id="command.database.prisma",
+            suffix="production-migrate",
+            title="Prisma production migration",
+            description="Identifies Prisma migration deployment against the selected database.",
+            matcher=_PRISMA_DEPLOY,
+            action_class="Prisma production migration command",
+            risk_classes=("execution", "network_egress"),
+            safer_alternative="Review pending migration files and verify the target database before deployment.",
+            safe_variants=help_variants(_PRISMA_DEPLOY, short=True),
+            example_command="npx prisma migrate deploy",
+        ),
+    rule(
+            extension_id="command.platform.firebase",
+            suffix="production-deploy",
+            title="Firebase production deployment",
+            description="Identifies deployment of Firebase application resources and security configuration.",
+            matcher=_FIREBASE_DEPLOY,
+            action_class="Firebase production command",
+            risk_classes=("execution", "network_egress"),
+            safer_alternative="Confirm the active Firebase project and deploy target before publishing.",
+            safe_variants=help_variants(_FIREBASE_DEPLOY, short=True),
+            example_command="firebase deploy",
+        ),
+)
