+++ src/codex_plugin_scanner/guard/runtime/command_common_cli_rules_2.py
@@ -1,0 +1,161 @@
+"""Common CLI command rules, batch 2."""
+
+from __future__ import annotations
+
+from .command_common_cli_matchers import (
+    _ARGO_DESTRUCTIVE,
+    _ARGO_RECONCILE,
+    _BQ_DESTRUCTIVE,
+    _DOTNET_MUTATION,
+    _DOTNET_PUBLICATION,
+    _FIREBASE_DESTRUCTIVE,
+    _FLUX_DESTRUCTIVE,
+    _FLUX_RECONCILE,
+    _FLY_CHANGE,
+    _FLY_DESTRUCTIVE,
+    _RAILWAY_DESTRUCTIVE,
+)
+from .command_common_cli_rule_support import help_variants, rule
+from .command_rules import CommandSafetyRule
+
+COMMON_CLI_COMMAND_RULES_2: tuple[CommandSafetyRule, ...] = (
+    rule(
+            extension_id="command.platform.firebase",
+            suffix="destructive",
+            title="Firebase destructive operation",
+            description="Identifies deletion of functions, application data, or Hosting state.",
+            matcher=_FIREBASE_DESTRUCTIVE,
+            action_class="Firebase destructive command",
+            risk_classes=("destructive_shell", "network_egress"),
+            safer_alternative="Inspect the exact Firebase project, target, and recovery options before deletion.",
+            severity="critical",
+            safe_variants=help_variants(_FIREBASE_DESTRUCTIVE, short=True),
+            example_command="firebase functions:delete worker",
+        ),
+    rule(
+            extension_id="command.gitops.argocd",
+            suffix="destructive",
+            title="Argo CD destructive operation",
+            description="Identifies application, ApplicationSet, cluster, or repository removal through argocd.",
+            matcher=_ARGO_DESTRUCTIVE,
+            action_class="Argo CD destructive command",
+            risk_classes=("destructive_shell", "network_egress"),
+            safer_alternative="Inspect the selected application and cascade behavior before deletion.",
+            severity="critical",
+            safe_variants=help_variants(_ARGO_DESTRUCTIVE, short=True),
+            example_command="argocd app delete production",
+        ),
+    rule(
+            extension_id="command.gitops.argocd",
+            suffix="reconciliation",
+            title="Argo CD reconciliation mutation",
+            description="Identifies sync, rollback, patch, and operation termination through argocd.",
+            matcher=_ARGO_RECONCILE,
+            action_class="Argo CD reconciliation command",
+            risk_classes=("execution", "network_egress"),
+            safer_alternative="Inspect application diff and health before changing live reconciliation state.",
+            safe_variants=help_variants(_ARGO_RECONCILE, short=True),
+            example_command="argocd app sync production",
+        ),
+    rule(
+            extension_id="command.gitops.flux",
+            suffix="destructive",
+            title="Flux destructive operation",
+            description="Identifies uninstall and deletion of Flux reconciliation resources.",
+            matcher=_FLUX_DESTRUCTIVE,
+            action_class="Flux destructive command",
+            risk_classes=("destructive_shell", "network_egress"),
+            safer_alternative="Inspect ownership and prune behavior before deleting or uninstalling Flux resources.",
+            severity="critical",
+            safe_variants=help_variants(_FLUX_DESTRUCTIVE, short=True),
+            example_command="flux delete kustomization production",
+        ),
+    rule(
+            extension_id="command.gitops.flux",
+            suffix="reconciliation",
+            title="Flux reconciliation mutation",
+            description="Identifies reconcile, suspend, and resume operations that change live GitOps behavior.",
+            matcher=_FLUX_RECONCILE,
+            action_class="Flux reconciliation command",
+            risk_classes=("execution", "network_egress"),
+            safer_alternative="Inspect source revision and reconciliation status before forcing or suspending changes.",
+            safe_variants=help_variants(_FLUX_RECONCILE, short=True),
+            example_command="flux reconcile kustomization production",
+        ),
+    rule(
+            extension_id="command.package.dotnet",
+            suffix="package-mutation",
+            title=".NET package dependency mutation",
+            description="Identifies .NET and NuGet package addition, install, and restore operations.",
+            matcher=_DOTNET_MUTATION,
+            action_class=".NET package mutation command",
+            risk_classes=("supply_chain", "network_egress", "execution"),
+            safer_alternative="Use locked dependencies and inspect package provenance before changing dependencies.",
+            safe_variants=help_variants(_DOTNET_MUTATION, short=True),
+            example_command="dotnet package add Example.Package",
+        ),
+    rule(
+            extension_id="command.package.dotnet",
+            suffix="publication",
+            title="NuGet publication mutation",
+            description="Identifies push or deletion of NuGet package versions.",
+            matcher=_DOTNET_PUBLICATION,
+            action_class=".NET package publication command",
+            risk_classes=("supply_chain", "network_egress", "destructive_shell"),
+            safer_alternative="Verify package, version, feed, and signing before changing published state.",
+            severity="critical",
+            safe_variants=help_variants(_DOTNET_PUBLICATION, short=True),
+            example_command="dotnet nuget push package.nupkg",
+        ),
+    rule(
+            extension_id="command.database.bigquery",
+            suffix="destructive",
+            title="BigQuery destructive operation",
+            description="Identifies resource deletion and replace-enabled load operations through bq.",
+            matcher=_BQ_DESTRUCTIVE,
+            action_class="BigQuery destructive command",
+            risk_classes=("destructive_shell", "network_egress"),
+            safer_alternative="Inspect the exact project, dataset, table, and destination before destructive changes.",
+            severity="critical",
+            safe_variants=help_variants(_BQ_DESTRUCTIVE, short=True),
+            example_command="bq rm -r -f project:dataset",
+        ),
+    rule(
+            extension_id="command.platform.fly",
+            suffix="destructive",
+            title="Fly.io destructive operation",
+            description="Identifies app, Machine, volume, or IP destruction through flyctl.",
+            matcher=_FLY_DESTRUCTIVE,
+            action_class="Fly.io destructive command",
+            risk_classes=("destructive_shell", "network_egress"),
+            safer_alternative="Inspect the exact app and dependent resource before destruction.",
+            severity="critical",
+            safe_variants=help_variants(_FLY_DESTRUCTIVE, short=True),
+            example_command="fly apps destroy production",
+        ),
+    rule(
+            extension_id="command.platform.fly",
+            suffix="production-change",
+            title="Fly.io deployment or secret change",
+            description="Identifies Fly.io deploy and application secret mutation operations.",
+            matcher=_FLY_CHANGE,
+            action_class="Fly.io production command",
+            risk_classes=("execution", "network_egress"),
+            safer_alternative="Confirm app, organization, deployment, and secret scope before remote changes.",
+            safe_variants=help_variants(_FLY_CHANGE, short=True),
+            example_command="fly deploy",
+        ),
+    rule(
+            extension_id="command.platform.railway",
+            suffix="destructive",
+            title="Railway destructive operation",
+            description="Identifies project or volume deletion through Railway CLI.",
+            matcher=_RAILWAY_DESTRUCTIVE,
+            action_class="Railway destructive command",
+            risk_classes=("destructive_shell", "network_egress"),
+            safer_alternative="Confirm project, service, environment, and persistent data before deletion.",
+            severity="critical",
+            safe_variants=help_variants(_RAILWAY_DESTRUCTIVE, short=True),
+            example_command="railway delete --yes",
+        ),
+)
